refactor setting utils under util package
